backend/tasks/requirements.txt
asgiref==3.8.1
crewai==0.32.2
crewai-tools==0.3.0
Django==5.0.7
djangorestframework==3.15.2
sqlparse==0.5.1
python-dotenv>=1.0.1,<1.1
psycopg2-binary>=2.9.9,<2.10
django-cors-headers==4.4.0